requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-aws
bs4
tiktoken
boto3
chromadb
awscli
streamlit
pypdf
faiss-cpu
transformers
accelerate
bitsandbytes
python-dotenv
sentencepiece
ctransformers
-e .